chore: rename EntityPanelUploader to EntityPanelUploader
