+++ shared/src/components/EntityPanelUploader/EntityPanelUploader.styled.tsx
