+++ shared/src/components/index.ts
@@ -7,7 +7,7 @@ export * from './ReviewableCard'
 export * from './ReviewableProgressCard'
 export * from './FileThumbnail'
 export * from './Thumbnail'
-export * from './EntityThumbnailUploader'
+export * from './EntityPanelUploader'
 export * from './ThumbnailUploader'
 export * from './SimpleFormDialog'
 export * from './EntityPath'
